+++ CHANGELOG.md
@@ -20,10 +20,13 @@
 
     - mark `binary-search` as `dontCheck`
     - mark `haskoin-core` as `dontCheck`
-    - mark `http-client`, `http-client-openssl`, and `http-client-tls` as `dontCheck`.
-    - mark `js-query` as `dontCheck`.
+    - mark `http-client` as `dontCheck`
+    - mark `http-client-openssl` as `dontCheck`
+    - mark `http-client-tls` as `dontCheck`
+    - mark `http-conduit` as `dontCheck`
+    - mark `js-query` as `dontCheck`
     - mark `prettyprinter` as `dontCheck`
-    - mark `servant-openapi3` as `dontCheck`.
+    - mark `servant-openapi3` as `dontCheck`
 
     Added in [#59](https://github.com/cdepillabout/stacklock2nix/pull/59).